% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateTiTV.R
\name{calculateTiTv}
\alias{calculateTiTv}
\title{A function which calculated the Ti/Tv ratio for a sample of vcf entries
given a vcfObj.}
\usage{
calculateTiTv(vcfObj)
}
\arguments{
\item{vcfObj}{A VcfObj S4 object (produced by loadVCF)}
}
\value{
a summary table of counts and ratios.
}
\description{
A function which calculates the Transition/Transversion ratio
for a sample of vcf entries given a vcfObj
}
\examples{
vcfData <- loadVCF(system.file("extdata", "text000.vcf", package="SNPVisualizR"))
calculateTiTv(vcfData)

}
\references{
Wang, Jing et al. “Genome measures used for quality control are dependent on
gene function and ancestry.” Bioinformatics (Oxford, England) vol. 31,3
(2015): 318-23. doi:10.1093/bioinformatics/btu668
}
