% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faClass.R
\docType{class}
\name{faObj-class}
\alias{faObj-class}
\title{faObj}
\description{
An S4 object which stores fasta file sequence data for use in the
functions as a part of this package.
}
\details{
An S4 class to represent the contents of a FASTA file.
}
\section{Slots}{

\describe{
\item{\code{sequenceData}}{A data table containing the sequence data for the fasta
sequences.}

\item{\code{codonData}}{A data table containing the formatted codon data for the
sequence

@importFrom data.table data.table}
}}

