% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadFA.R
\docType{data}
\name{loadFA}
\alias{loadFA}
\alias{filepath}
\title{Loads the contents of the fasta file and returns a S4 object containing the
sequence information as a data frame}
\format{
An object of class \code{NULL} of length 0.
}
\usage{
filepath
}
\arguments{
\item{filepath}{A string path to the vcf file you would like to load. Please
ensure that your file meets {specification something}}
}
\value{
an faObj containing the sequence data.
}
\description{
Loads the contents of the fasta file and returns a S4 object containing the
sequence information as a data frame
}
\examples{
#fafile contains the relative path the file.
fafile <- system.file("extdata", "test01.fa", package="SNPVisualizR")
faObj <- loadFA(fafile)
}
\keyword{datasets}
