% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadvcf.R
\name{loadVCF}
\alias{loadVCF}
\title{loadVCF loads a vcf file's contents into a VCF object for use with the other
functions within this package.}
\usage{
loadVCF(filepath, fileSizeLimit = 1e+05, annotated = FALSE)
}
\arguments{
\item{filepath}{A string path to the vcf file you would like to load. Please
ensure that your file meets the 1000Genomes VCF v4.0 file specification.}

\item{fileSizeLimit}{number of lines to process (useful with very
large files) (1E06 by default).}

\item{annotated}{A logical stating whether or not your vcf file has been
annotated.}
}
\value{
a vcfObj with the corresponding species name, contig data and
variant data
}
\description{
loadVCF loads a vcf file's contents into a VCF object for use with the other
functions within this package.
}
