% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPosition.R
\name{plotPosition}
\alias{plotPosition}
\title{Function which plots the variants by their position on the chromosome.}
\usage{
plotPosition(
  vcfInput,
  faInput,
  sections = 300,
  colorBy = "chrom",
  divideBy = "chrom"
)
}
\arguments{
\item{vcfInput}{A VCFObj S4 object which contains variant entries.}

\item{faInput}{Optional; a S4 object which contains the fa sequences.}

\item{sections}{number of sections to include in the position plot ()}

\item{colorBy}{color variants by filter, showing percent variants'
postion colored by filter}

\item{divideBy}{group variant entries and show position by group}
}
\description{
Function which plots the variants by their position on the chromosome.
}
\references{

}
