% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateTiTV.R
\docType{data}
\name{ref_a}
\alias{ref_a}
\title{A function which calculates the Transition/Transversion ration for a sample
of vcf entries given a vcfObj}
\format{
An object of class \code{NULL} of length 0.
}
\usage{
ref_a
}
\arguments{
\item{vcfObj}{A VcfObj S4 object (produced by loadVCF)}
}
\value{
a summary table of counts and ratios.
}
\description{
A function which calculates the Transition/Transversion ration for a sample
of vcf entries given a vcfObj
}
\keyword{datasets}
