% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runSNPVisualizR.R
\name{runSNPVisualizR}
\alias{runSNPVisualizR}
\title{Launches the corresponding SNPVisualizR shiny app.}
\usage{
runSNPVisualizR()
}
\value{
Nothing; only runs shiny app.
}
\description{
Function which launches the SNPVisualizR
}
\examples{
\dontrun{
    runSNPVisualizR
}
}
