% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variantClass.R
\docType{class}
\name{variantEntry-class}
\alias{variantEntry-class}
\title{An S4 class which represents a variant entry (within a vcf file).
See 1000Genome's VCF4.2 definition (samtools github for additional information on 4.2)}
\description{
An S4 class which represents a variant entry (within a vcf file).
See 1000Genome's VCF4.2 definition (samtools github for additional information on 4.2)
}
\section{Slots}{

\describe{
\item{\code{species}}{The VCF file's corresponding species.}

\item{\code{contigInfo}}{A list containing objects with the contig info.}

\item{\code{variantData}}{A list containing variant data (reference|alt)}

\item{\code{id}}{The ID of the variant's entry.}

\item{\code{ref_a}}{The Reference Allele}

\item{\code{alt_a}}{The alternate Allele.}

\item{\code{qual}}{The quality of the variant}

\item{\code{filter}}{whether the variant passes any filters that have been}
}}

