% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcfClass.R
\name{vcfObj}
\alias{vcfObj}
\title{An S4 class to represent a vcf file's variant data.}
\description{
An S4 object which stores the vcf entries for usee in the
functions as a part of this package.
}
\section{Slots}{

\describe{
\item{\code{species}}{The VCF file's corresponding species.}

\item{\code{contigInfo}}{A vector containing objects with the contig info.}

\item{\code{variantData}}{A data table containing the vcf entries.}
}}

