% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faClass.R
\docType{class}
\name{faObj-class}
\alias{faObj-class}
\title{An S4 class to represent the contents of a FASTA file.}
\description{
An S4 object which stores fasta file sequence data for use in the
functions as a part of this package.
}
\section{Slots}{

\describe{
\item{\code{sequenceData}}{A data table containing the sequence data for the fasta
sequences.

@import data.table}
}}

