% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPosition.R
\name{plotPosition}
\alias{plotPosition}
\title{Function which plots the variants by their position on the chromosome.}
\usage{
plotPosition(vcfInput, sections = 300, colorBy = "chrom", divideBy = "chrom")
}
\arguments{
\item{vcfInput}{A VCFObj S4 object which contains variant entries.}

\item{sections}{number of sections to include in the position plot
(wrapper parameter for ggplot2's "bins")}

\item{colorBy}{color variants by filter, showing percent variants'
postion colored by filter}

\item{divideBy}{group variant entries and show position by group}
}
\description{
Function which plots the variants by their position on the chromosome.
}
\examples{
# First, load vcf data into variable.
getwd()
vcfData <- loadVCF(
  system.file("extdata", 'text000.vcf', package = "SNPVisualizR")
)
plotPosition(vcfData)

}
