% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateTiTV.R
\docType{data}
\name{ref_a}
\alias{ref_a}
\title{A function which calculates the Transition/Transversion ration for a sample
of vcf entries given a vcfObj}
\format{
An object of class \code{NULL} of length 0.
}
\usage{
ref_a
}
\arguments{
\item{vcfObj}{A VcfObj S4 object (produced by loadVCF)}
}
\value{
a summary table of counts and ratios.
}
\description{
A function which calculates the Transition/Transversion ration for a sample
of vcf entries given a vcfObj
}
\examples{
vcfData <- loadVCF("inst/examples/text000.vcf")
calculateTiTv(vcfData)

}
\references{
Wang, Jing et al. “Genome measures used for quality control are dependent on
gene function and ancestry.” Bioinformatics (Oxford, England) vol. 31,3
(2015): 318-23. doi:10.1093/bioinformatics/btu668
}
\keyword{datasets}
